### 2019-08-02 #1624 @labarba
👋 @bwoshea — something is not quite right with `paper.md`…not sure what. I do notice that the indentation of the tags doesn't quite match the other blocks. Could this be it?

Also, JOSS papers usually reside in the software repository (a different branch is OK), but here you've created a separate repo for just the paper. Is this the software repo?
https://github.com/enzo-project/enzo-dev

Note that in the JOSS workflow the software repository gets added to the decorators on the first-page margin of the published paper, so it kind of breaks things if you put the paper in another repo.
