### 2019-08-02 #1624 @danielskatz
Hi @bwoshea - suggestions for potential reviewers are very welcome. Please consider people on [this list](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
