### 2019-08-02 #1624 @danielskatz
@bwoshea - part of the problem with the paper compiling is (or will be) blank ORCIDs - for authors without ORCIDS, remove the blank ORCID line from the .md file
