### 2019-08-03 #1624 @labarba
Oh, yes, absolutely! We've defaulted to a minimum of two reviewers in JOSS now. But maybe here we need three? Do you know anyone you might recommend?
