### 2019-08-03 #1624 @labarba
👋  @bwoshea — to stave off any confusion: I am the Associate-Editor-in-Chief on rotation this week, and generally pushing things along for new submissions (or stuck ones). @danielskatz volunteered to be your handling editor, and will take this into review.
