### 2019-08-03 #1624 @danielskatz
👋 @r-j-arnold - please let me know if you would be willing to review this for JOSS
