### 2019-08-04 #1624 @bwoshea
@labarba thank you for all of the information!  I just made some updates to the repository and will test it momentarily.

Regarding repositories, would it be fine if we moved the paper to a subdirectory called 'method-paper-JOSS-2019' within our code repository?  It's a bit unclear from the JOSS submission instructions.  I can do that essentially immediately.
