### 2019-08-04 #1624 @labarba
> would it be fine if we moved the paper to a subdirectory called 'method-paper-JOSS-2019' within our code repository?

You could do that. Our bot [`@whedon`](https://joss.readthedocs.io/en/latest/whedon.html) knows how to look into folders for the paper source. You can also put the paper in a branch, in which case the command to compile will be `@whedon generate pdf from branch <branch-name>`. In any case, we will have to change the repository listed here by editing the top comment. Let us know the URL here.
