### 2019-08-04 #1624 @labarba
@arfon — have a look at this compiled [article proof](https://github.com/openjournals/joss-papers/blob/joss.01624/joss.01624/10.21105.joss.01624.pdf). With a very long author list, the margin decorators get pushed to the bottom of the page, colliding with the footer. This may require some change to the paper template.
