### 2019-08-04 #1624 @bwoshea
@labarba @arfon I've made a few updates to the repo to fix a few formatting issues, but the link @labarba shared a few minutes ago seems to point to the correct PDF.  Thanks to all for the suggestions about how to sort out the compilation issue!
