### 2019-08-04 #1624 @labarba
> Would it be OK to proceed with the review using the current repository for the moment?

I believe so.
@arfon — does manually editing the URL to the software repository on the top comment of the Review issue update the link on the paper's front page in subsequent compilations?
