### 2019-08-04 #1624 @arfon
> @arfon — does manually editing the URL to the software repository on the top comment of the Review issue update the link on the paper's front page in subsequent compilations?

It does not but I can fix up the JOSS site so that it will.
