### 2019-08-04 #1624 @bwoshea
> > @arfon — does manually editing the URL to the software repository on the top comment of the Review issue update the link on the paper's front page in subsequent compilations?
>
> It does not but I can fix up the JOSS site so that it will.

Excellent, thanks!  I just messaged all of the Enzo developers, and we'll move on the repository issue pretty quickly as soon as we reach consensus (meaning on the 2-3 day time frame, given the number of developers).
