### 2019-08-04 #1624 @arfon
> @arfon — have a look at this compiled article proof. With a very long author list, the margin decorators get pushed to the bottom of the page, colliding with the footer. This may require some change to the paper template.

Thanks for spotting this. This is something we'll have to fix at the end of the review as I'll have to manually override the margin formatting for this paper.
