### 2019-08-05 230986a
Switch configure script to bash.
