### 2019-08-05 d17e5f4
Keep csh installed for tests.
