### 2019-08-05 feee1c2
Add test of various compile options.
