### 2019-08-05 439b5c5
Bug fixes required for non-default precision compilation.
