### 2019-08-06 230949f
Skip loading test-suite config and start from default settings.
