### 2019-08-06 e98b203
Output compile config if build fails.
