+++ .circleci/config.yml
@@ -119,7 +119,7 @@ commands:
           make precision-$prec
           make particles-$part
           make clean
-          make -j 4
+          make -j 4 || make show-config
           done ; done
           # return this to default
           make precision-64
@@ -135,12 +135,12 @@ commands:
           make integers-$inte
           make particle-id-$pids
           make clean
-          make -j 4
+          make -j 4 || make show-config
           done ; done
           # test without mpi
           make use-mpi-no
           make clean
-          make -j 4
+          make -j 4 || make show-config
 
   build-docs:
     description: "Test the docs build."
