### 2019-08-06 50ab756
Make sure to return error if build fails.
