+++ .circleci/config.yml
@@ -105,7 +105,6 @@ commands:
       - run: |
           source $BASH_ENV
           source $HOME/venv/bin/activate
-          alias build-enzo='make -j 4 || (make show-config ; exit 1)'
           ./configure
           cd src/enzo
           make machine-linux-gnu
@@ -120,7 +119,7 @@ commands:
           make precision-$prec
           make particles-$part
           make clean
-          build-enzo
+          'make -j 4 || (make show-config ; exit 1)'
           done ; done
           # return this to default
           make precision-64
@@ -136,12 +135,12 @@ commands:
           make integers-$inte
           make particle-id-$pids
           make clean
-          build-enzo
+          'make -j 4 || (make show-config ; exit 1)'
           done ; done
           # test without mpi
           make use-mpi-no
           make clean
-          build-enzo
+          'make -j 4 || (make show-config ; exit 1)'
 
   build-docs:
     description: "Test the docs build."
