### 2019-08-06 10c17a4
Fix three warnings.
