+++ src/enzo/Grid.h
@@ -635,7 +635,7 @@ gradient force to gravitational force for one-zone collapse test. */
 
 /* Debugging support. */
 
-   int DebugCheck(char *message = "Debug");
+   int DebugCheck(const char *message = "Debug");
 
 #ifdef EMISSIVITY
    /* define function prototype as a grid member function */
