+++ src/enzo/WriteHDF5HierarchyFile.C
@@ -32,9 +32,7 @@ int SetGlobalGridID(int &GlobalID, HierarchyEntry *Grid);
 // the following HDF5 helper routines are defined in
 // Grid_WriteHierarchyInformationHDF5.C
 int HDF5_WriteAttribute(hid_t group_id, const char *AttributeName, int Attribute, FILE *log_fptr);
-#ifdef SMALL_INTS
-int HDF5_WriteAttribute(hid_t group_id, const char *AttributeName, Eint64 Attribute, FILE *log_fptr);
-#endif
+int HDF5_WriteAttribute(hid_t group_id, const char *AttributeName, PINT Attribute, FILE *log_fptr);
 int HDF5_WriteAttribute(hid_t group_id, const char *AttributeName, FLOAT Attribute, FILE *log_fptr);
 int HDF5_WriteDataset(hid_t group_id, const char *DatasetName, int *Dataset, int NumberOfElements, FILE *log_fptr);
 
