### 2019-08-06 6b88936
Add test for compiling with photon-no and fix one issue.
