+++ .circleci/config.yml
@@ -137,8 +137,15 @@ commands:
           make clean
           make -j 4 || (make show-config ; exit 1)
           done ; done
+          make integers-64
+          make particle-id-64
           # test without mpi
           make use-mpi-no
+          make clean
+          make -j 4 || (make show-config ; exit 1)
+          make use-mpi-yes
+          # test ray-tracing
+          make use-photon-no
           make clean
           make -j 4 || (make show-config ; exit 1)
 
