### 2019-08-06 ef4a804
Fix photon compile flag.
