+++ method-paper-joss-2019/paper.bib
@@ -1,0 +1,506 @@
+@online{EnzoGitRepo,
+  author = {Enzo Developers},
+  title = {Enzo GitHub repository},
+  year = 2019,
+  url = {https://github.com/enzo-project/enzo-dev},
+  urldate = {2019-07-19}
+}
+
+@online{EnzoReleaseNotes,
+  author = {Enzo Developers},
+  title = {Enzo Release Notes},
+  year = 2019,
+  url = {https://enzo-project.org/ReleaseNotes.html},
+  urldate = {2019-07-19}
+}
+
+@ARTICLE{EnzoMethodPaper2014,
+       author = {{Bryan}, Greg L. and {Norman}, Michael L. and {O'Shea}, Brian W. and
+         {Abel}, Tom and {Wise}, John H. and {Turk}, Matthew J. and
+         {Reynolds}, Daniel R. and {Collins}, David C. and {Wang}, Peng and
+         {Skillman}, Samuel W. and {Smith}, Britton and {Harkness}, Robert P. and
+         {Bordner}, James and {Kim}, Ji-hoon and {Kuhlen}, Michael and
+         {Xu}, Hao and {Goldbaum}, Nathan and {Hummels}, Cameron and
+         {Kritsuk}, Alexei G. and {Tasker}, Elizabeth and {Skory}, Stephen and
+         {Simpson}, Christine M. and {Hahn}, Oliver and {Oishi}, Jeffrey S. and
+         {So}, Geoffrey C. and {Zhao}, Fen and {Cen}, Renyue and {Li}, Yuan and
+         {Enzo Collaboration}},
+        title = "{ENZO: An Adaptive Mesh Refinement Code for Astrophysics}",
+      journal = {The Astrophysical Journal Supplements},
+     keywords = {hydrodynamics, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = "2014",
+        month = "Apr",
+       volume = {211},
+       number = {2},
+          eid = {19},
+        pages = {19},
+          doi = {10.1088/0067-0049/211/2/19},
+archivePrefix = {arXiv},
+       eprint = {1307.2265},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014ApJS..211...19B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{WiseNature2019,
+       author = {{Wise}, John H. and {Regan}, John A. and {O'Shea}, Brian W. and
+         {Norman}, Michael L. and {Downes}, Turlough P. and {Xu}, Hao},
+        title = "{Formation of massive black holes in rapidly growing pre-galactic gas clouds}",
+      journal = {\nat},
+     keywords = {Astrophysics - Astrophysics of Galaxies},
+         year = "2019",
+        month = "Jan",
+       volume = {566},
+       number = {7742},
+        pages = {85-88},
+          doi = {10.1038/s41586-019-0873-4},
+archivePrefix = {arXiv},
+       eprint = {1901.07563},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019Natur.566...85W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{SmithPop2Prime2015,
+       author = {{Smith}, Britton D. and {Wise}, John H. and {O'Shea}, Brian W. and
+         {Norman}, Michael L. and {Khochfar}, Sadegh},
+        title = "{The first Population II stars formed in externally enriched mini-haloes}",
+      journal = {\mnras},
[…]
