### 2019-08-09 #1624 @bwoshea
@danielskatz - yes, I'm moving it over to the enzo-dev repository right now.  As soon as the PR is approved you will be able to find it in the master branch of github.com/enzo-project/enzo-dev in the subdirectory "method-paper-joss-2019".  I'm working to get our required number of approvals right now, and expect that will happen pretty soon.
