### 2019-08-09 #1624 @bwoshea
@danielskatz - the PR is now merged, and you can find the paper in the Enzo code repository (specifically, in https://github.com/enzo-project/enzo-dev/tree/master/method-paper-joss-2019) .  Do you need anything else to start the review?
