### 2019-08-09 #1624 @labarba
I changed the repository URL on the top comment of this issue, then compiled with `whedon`, and it's all good!
