+++ method-paper-joss-2019/paper.bib
@@ -1,5 +1,5 @@
 @online{EnzoGitRepo,
-  author = {Enzo Developers},
+  author = {{Enzo Developers}},
   title = {Enzo GitHub repository},
   year = 2019,
   url = {https://github.com/enzo-project/enzo-dev},
@@ -7,7 +7,7 @@
 }
 
 @online{EnzoReleaseNotes,
-  author = {Enzo Developers},
+  author = {{Enzo Developers}},
   title = {Enzo Release Notes},
   year = 2019,
   url = {https://enzo-project.org/ReleaseNotes.html},
