### 2019-08-09 #1636 @danielskatz
👋 @zingale, @rtfisher - we are ready for you to review this now.  Please read the first 2 comments in this issue carefully.  Your job now is to go through the paper (the latest "check article proof" in this issue) and the repository, and check off items in your checklist until you are done, and if you find problems, either report them here or create an issue in the repo and mention this issue in that one.

If you have any questions or problems, please let me know.
