### 2019-08-09 588d114
Updated mechanical stars implementation
