+++ src/enzo/Grid.h
@@ -2879,6 +2879,19 @@ int zEulerSweep(int j, int NumberOfSubgrids, fluxes *SubgridFluxes[],
 
   int ApplySmartStarParticleFeedback(ActiveParticleType** ThisParticle);
 
+
+int MechStars_Creation(grid* ParticleArray, float* Temperature, 
+        float *DMField, int level, float* CoolingTime, 
+        int MaximumNumberOfNewParticles, int* NumberOfParticlesSoFar);
+int MechStars_FeedbackRoutine(int level, float* mu_field);
+int MechStars_DepositFeedback(float supernovaEnergy, 
+                        float ejectaMass, float ejectaMetal,
+                        float* up, float* vp, float* wp,
+                        float* xp, float* yp, float* zp,
+                        int ip, int jp, int kp,
+                        int size, float* mu_field, int winds);
+
+
 //------------------------------------------------------------------------
 // Radiative transfer methods that don't fit in the TRANSFER define
 //------------------------------------------------------------------------
