### 2019-08-10 0eefcb7
Corrected error in rates determination
