### 2019-08-11 #1636 @danielskatz
> @bwoshea - merging [enzo-project/enzo-dev#93](https://github.com/enzo-project/enzo-dev/pull/93) will fix 2 small bib entries

Now that this has been done, I will recompile - @bwoshea, you can do this too, just the way I am, with an instruction in whedon in a new comment (letting you know in case more changes are needed)
