### 2019-08-11 #1636 @zingale
regarding

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)

There are 2 references to Enzo itself, referring to github.  Should these be changed to an archived version of the code with a DOI?
