### 2019-08-11 #1636 @zingale
The license is UI/NCSA until 2000 and UC BSD until 2008, both are on the OSI list.  Should we assume the last license remains in effect since 2008?
