### 2019-08-11 enzo-project/enzo-dev#94 @zingale
[JOSS REVIEW] LICENSE
The license is University of Illinois/NCSA Open Source License until 2000 and UC BSD from 2000-2008.  Should we assume that that license remains in effect today?  Should the year be extended?
