### 2019-08-11 enzo-project/enzo-dev#95 @zingale
[JOSS REVIEW] user's guide intro
In the User's Guide intro,

https://enzo.readthedocs.io/en/enzo-2.6/user_guide/index.html

It says:

> Arguably the most useful and important piece of this guide is Enzo Parameter List, which contains descriptions of all of the roughly 300 possible input parameters (as of September 2008).

Have there been no new papers in over a decade? or is the list out of date? or is that statement out of date?

It also says:

> Detailed information on the algorithms used in Enzo will be available in the method paper (unreleased as of September 2008). In the meantime, see the Enzo Primary References for more concrete Enzo information.

Is that still true? or is the 2014 paper referenced in the JOSS paper the method paper this is referring to?
