### 2019-08-11 enzo-project/enzo-dev#96 @zingale
[JOSS REVIEW] linux-gnu build info is distro specific
Upon running
```
make machine-linux-gnu
```
it outputs:
```
 *** Execute 'gmake clean' before rebuilding executables ***

   MACHINE: Use apt-get to install libhdf5-serial-dev gfortran openmpi-bin libopenmpi-dev
   MACHINE-NAME: linux-gnu
```

but many Linux distros do not use `apt-get` and new users could be confused.  It would be helpful to have a list of the requirements needed for Enzo in the README.md or a REQUIREMENTS.txt
