### 2019-08-11 #1636 @zingale
https://github.com/enzo-project/enzo-dev/issues/96
