### 2019-08-11 enzo-project/enzo-dev#97 @zingale
[JOSS REVIEW] documentation link in README.md broken
The `README.md` lists:

 http://enzo-project.org/docs/2.6/

for the docs, but this link is broken and gives a 404 error
