### 2019-08-11 enzo-project/enzo-dev#98 @zingale
[JOSS REVIEW] quick start references to yt are inaccurate
The quick start page:

https://enzo.readthedocs.io/en/enzo-2.6/user_guide/bootcamp.html

starts by saying:

```
This is an unofficial quickstart guide to using Enzo, starting from checking out the Enzo and YT sources, through to building them, then to running an example problem and looking at its output.
```

but no reference to the yt sources is found on the page.  Further, it ends by saying:

```
At this point, you’ve (hopefully!) run an Enzo simulation. You should also have yt set up.
```

but there was no mention of helping setup yt.  This should be fixed or changed.

Also, the guide says "This is an unofficial quickstart ", but these are the "official" docs, so this is a bit confusing, and suggests that perhaps there is an official quickstart somewhere?
