### 2019-08-11 enzo-project/enzo-dev#94 @zingale
Somewhat related, the docs are copyright 2012:

https://enzo.readthedocs.io/en/enzo-2.6/user_guide/bootcamp.html

(see the footer).  Should that be updated too?
