### 2019-08-11 enzo-project/enzo-dev#99 @zingale
[JOSS REVIEW] quck start guide should show output
In the quick start guide, the user is told to visualize the output with yt.  To help understand if their run was successful, it would be helpful to have the images of the output inline in the quickstart guide to compare with.
