### 2019-08-11 enzo-project/enzo-dev#100 @zingale
[JOSS REVIEW] Hard to learn about AMR capabilities
Reading the documentation, I can't find basic information about how the AMR works in Enzo. The main reference is a link to Wikipedia on https://enzo.readthedocs.io/en/enzo-2.6/user_guide/index.html

 A natural place would be:

https://enzo.readthedocs.io/en/enzo-2.6/physics/index.html

In particular, it should be easy to find whether Enzo uses something like Berger and Colella AMR (or a subset), an oct-tree block structured AMR (like Flash/Paramesh, e.g.) or cell-by-cell AMR (like Rage).  Also, what refinement ratios are supported and whether time subcycling is used.  This desciption should be somewhere up front to help potential users understand the capabilities.
