### 2019-08-12 ac50d80
remove duplicate FAQ entry
