### 2019-08-13 enzo-project/enzo-dev#106 @zingale
[JOSS REVIEW] how to create a problem + list of distributed problems?
There are a lot of problem setups in `enzo-dev/run`, but I can't seem to find descriptions of them in the docs.  It would be good to have a list of all of these problems with a 1-2 sentence desciption of them.

Also, I can't seem to find instructions on how to add a problem to enzo.  For instance, if I wanted to add a Rayleigh-Taylor instability problem, it is not clear how I would start.
