### 2019-08-13 #1636 @zingale
https://github.com/enzo-project/enzo-dev/issues/106
