### 2019-08-14 #1636 @zingale
I'm done with my pass through the docs and have filed a number of issues, some which I would like to see addressed for the review (marked with `[JOSS REVIEW]`) and others that are suggestions to take into consideration.

I will try running some standard test problems next.
