### 2019-08-14 a15483f
Vast debugging from frontera. Get crashes with and without SMM.  Is it fronteras MPI? run on comet to see.
