+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -42,8 +42,8 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         and all have radius dx from the source particle. 
         Each vertex particle will then be CIC deposited to the grid!
     */
-    bool debug = true;
-    bool criticalDebug = true;
+    bool debug = false;
+    bool criticalDebug = false;
     int index = ip+jp*GridDimension[0]+kp*GridDimension[0]*GridDimension[1];
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
     
@@ -99,11 +99,11 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         of all quantities and be coupled to the grid after the cic deposition. */
     float density [size];
     float metals [size];
-    float *u = new float [size];
-    float *v = new float[size];
-    float *w = new float [size];
-    float *totalEnergy =new float [size];
-    float *gasEnergy =new float[size];
+    float u [size];
+    float v [size];
+    float w [size];
+    float totalEnergy  [size];
+    float gasEnergy [size];
     for (int i=0; i<size; i++){
         density[i] = 0.0;
         metals[i] = 0.0;
@@ -162,7 +162,7 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
                 CloudParticleVectorY[cpInd] /= norm;
                 CloudParticleVectorX[cpInd] /= norm;  
             }
-    /* Each particle gets 1/12 of energy, momenta, mass, and metal.  There
+    /* Each particle gets 1/32 of energy, momenta, mass, and metal.  There
     are no varying vector / scalar weights to worry about.  The momenta coupled
     is simply \hat(r_ba) p/12 for r_ba the vector from source to coupled
     particle.  */
@@ -243,10 +243,8 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
             usePt = 1;
             
         /* Determine coupled momenta if rc < dx 
-        else couple p_ej*(1+dx/r_cool)**4 */
+        else couple p_ejecta */
             if(debug)fprintf(stdout, "Using P_t with Nb = %f, E= %e",nmean, coupledEnergy/1e51);
-            float Efactor = 1.0;
-            if (dxRatio > 4) Efactor = coupledEnergy/ejectaEnergy*pow(dxRatio,3);
             coupledMomenta = 4.8e5*pow(nmean, -1.0/7.0)
                 * pow(ejectaEnergy/1e51, 13.0/14.0) * fz; //Msun*km/s
         // }
@@ -272,7 +270,6 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         Analytic SNR shell mass is on, adjust the shell mass 
         Shell is limited on upper end by 1/1000 mass of 
             cell with mean density*/
-    float maxShellMass = 1.0*DensityUnits*MassUnits/1000; // large shell mass evacuates too efficently... muField-> 0.0 and NaN ensues!
     if (dxRatio <= 50 && dxRatio >= 0.1 && coupledEnergy > 0
         && AnalyticSNRShellMass){
             shellVelocity = 413.0 *pow(nmean, 1.0/7.0)
@@ -284,7 +281,7 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
                 can only contribute to the mass if the shell velocity is 
                 higher than the gas velocity.*/
             if (shellVelocity > vmean){ 
-                shellMass = max(1e3, coupledMomenta/shellVelocity); //Msun
[…]
