+++ src/enzo/Grid_MechStarsFeedbackRoutine.C
@@ -1,5 +1,5 @@
 /*
-    Routine to determine feedback quantities and couple them to the 
+    Routine to determine feedback quantities and couple them to the
     Grid.  Coupling follows the implementation of Hopkins 2017 with
     modification for Enzo's fixed grid
 
@@ -22,16 +22,16 @@
 #include "phys_constants.h"
 
 
-    int determineSN(float age, int* nSNII, int* nSNIA, float massMsun, 
+    int determineSN(float age, int* nSNII, int* nSNIA, float massMsun,
                     float TimeUnits, float dtFixed);
     int determineWinds(float age, float* eWinds, float* zWinds, float* mWinds,
                         float massMsun, float zZsun, float TimeUnits, float dtFixed);
     int checkCreationCriteria(float* Density, float* Metals,
                         float* Temperature,float* DMField,
-                        float* Vel1, float* Vel2, float* Vel3, 
+                        float* Vel1, float* Vel2, float* Vel3,
                         float* CoolingTime, int GridDim,
-                        float* shieldedFraction, float* freeFallTime, 
-                        float* dynamicalTime, int i, int j, int k, 
+                        float* shieldedFraction, float* freeFallTime,
+                        float* dynamicalTime, int i, int j, int k,
                         float Time, float* RefinementField, float CellWidth);
     int FindField(int field, int farray[], int numfields);
     int GetUnits(float *DensityUnits, float *LengthUnits,
@@ -45,16 +45,16 @@
 int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
 {
 
-    // fprintf(stdout,"IN FEEDBACK ROUTINE\n  %d   %d   %d\n", 
+    // fprintf(stdout,"IN FEEDBACK ROUTINE\n  %d   %d   %d\n",
         // SingleSN, StellarWinds, UnrestrictedSN);
     float stretchFactor = 1.4;//1/sqrt(2) to cover cell diagonal
     /* Get units to use */
 
     int dim, i, j, k, index, size, field, GhostZones = NumberOfGhostZones;
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
-    
+
     /* Compute size (in floats) of the current grid. */
-    
+
     size = 1;
     for (dim = 0; dim < GridRank; dim++)
         size *= GridDimension[dim];
@@ -73,11 +73,11 @@ int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
     if (GetUnits(&DensityUnits, &LengthUnits, &TemperatureUnits,
             &TimeUnits, &VelocityUnits, &MassUnits, this->Time) == FAIL) {
         fprintf(stderr, "Error in GetUnits.\n");
-        return FAIL;    
-    } 
-    double dx = CellWidth[0][0];    
+        return FAIL;
+    }
+    double dx = CellWidth[0][0];
     MassUnits = DensityUnits*pow(LengthUnits*dx, 3)/SolarMass;
-    /* 
+    /*
         get metallicity field and set flag; assumed true thoughout feedback
         since so many quantities are metallicity dependent
      */
@@ -87,7 +87,7 @@ int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
         MetallicityField = TRUE;
     else
         MetalNum = 0;
-    
+
     int numSN = 0;
 
     /* Begin Iteration of all particles */
@@ -103,7 +103,7 @@ int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
         if (ParticleType[pIndex] == PARTICLE_TYPE_STAR
                 && ParticleMass[pIndex] > 0.0
                 && ParticleAttribute[0][pIndex] > 0.0){
-                
+
             // if (StarMakerAgeCutoff)
             //     if ((Time-ParticleAttribute[0][pIndex])
[…]

[…]
