### 2019-08-15 #1636 @danielskatz
@zingale  - thanks for all your comments so far. If possible, when you open an issue in https://github.com/enzo-project/enzo-dev, just mention this review thread (https://github.com/openjournals/joss-reviews/issues/1636) in that issue rather than posted that issue in this thread.  This will insert a note here that will show if the issue is open or closed.
