### 2019-08-16 enzo-project/enzo-dev#94 @zingale
JOSS review: https://github.com/openjournals/joss-reviews/issues/1636
