### 2019-08-16 enzo-project/enzo-dev#98 @zingale
JOSS Review: https://github.com/openjournals/joss-reviews/issues/1636
