### 2019-08-17 fa9a971
reverted energy deposition to as defined in Hopkins. Tests were failures and the deposition matches that of star_maker3mom.F now as well
