### 2019-08-18 d1cd108
Reworked momentum decision tree to cover different phases of SNR
