+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -42,13 +42,13 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         and all have radius dx from the source particle. 
         Each vertex particle will then be CIC deposited to the grid!
     */
-    bool debug = false;
-    bool criticalDebug = false;
+    bool debug = true;
+    bool criticalDebug = true;
     int index = ip+jp*GridDimension[0]+kp*GridDimension[0]*GridDimension[1];
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
     
     /* Compute size (in floats) of the current grid. */
-    float stretchFactor =1.;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
+    float stretchFactor =1.25;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
                                 // in units of dx. Since the cloud forms a sphere shell, stretchFactor > 1 is not recommended
     size = 1;
     for (int dim = 0; dim < GridRank; dim++)
@@ -228,24 +228,61 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
 
     if (debug)fprintf(stdout, "Dx [pc] = %f\n", dx*LengthUnits/pc_cm);
     float dxRatio = stretchFactor*dx*LengthUnits/pc_cm/CoolingRadius;
+    float pEjectMod = pow(2.0*ejectaEnergy*(ejectaMass*SolarMass), 0.5)/SolarMass/1e5;
+    /* We want to couple one of four phases: free expansion, Sedov-taylor, shell formation, or terminal 
+    The first three phases are take forms from Kim & Ostriker 2015, the last from Cioffi 1988*/
+    float cellwidth = dx*LengthUnits/pc_cm;
+    float p_free = 0.0;//sqrt(ejectaMass*SolarMass*ejectaEnergy)/SolarMass/1e5;//1.73e4*sqrt(ejectaMass*ejectaEnergy/1e51/3.); // free exp. momentum eq 15
+    float r_free = 2.75*pow(ejectaMass/3/nmean, 1./3.); // free exp radius eq 2
+    
+    
+    // assuming r_sedov == dx, solve for t3
+    float t3_sedov = pow( cellwidth*pc_cm
+            /(5.0*pc_cm*pow(ejectaEnergy/1e51/nmean, 1.0/5.0)), 5./2.); 
+    float p_sedov = 2.21e4*pow(ejectaEnergy/1e51, 4./5.)
+            * pow(nmean, 1./5.)* pow(t3_sedov, 3./5.); // eq 16
+
+      // shell formation radius eq 8
+    float r_shellform = 22.6*pow(ejectaEnergy/1e51, 0.29)*pow(nmean, -0.42); 
+    // p_sf = m_sf*v_sf eq 9,11
+    float p_shellform = 3.3936e5*pow(ejectaEnergy/1e51, 0.94)*pow(nmean, -0.13) ; // p_sf = m_sf*v_sf eq 9,11
+  
+    /* termninal momentum */
+    float pTerminal = 4.8e5*pow(nmean, -1.0/7.0)
+                * pow(ejectaEnergy/1e51, 13.0/14.0) * fz; // cioffi 1988, as written in Hopkins 2018
 
     float coupledMomenta = 0.0;
     float eKinetic = 0.0;
-    /* termninal momentum */
-    float pTerminal = 4.8e5*pow(nmean, -1.0/7.0)
-                * pow(ejectaEnergy/1e51, 13.0/14.0) * fz;
-    /* analytic momentum for resolved cooling radii */
-    float pEjectMod = pow(2.0*ejectaEnergy*(ejectaMass*SolarMass), 0.5) 
[…]
