### 2019-08-19 81ca203
Enforced smooth momentum functional forms
