### 2019-08-22 enzo-project/enzo-dev#97 @rtfisher
I confirm this is still an issue.
