### 2019-08-22 enzo-project/enzo-dev#108 @rtfisher
[JOSS REVIEW] paper.md has empty references section
The paper.md has an empty references section.

https://github.com/enzo-project/enzo-dev/blob/master/method-paper-joss-2019/paper.md

JOSS Review: openjournals/joss-reviews#1636
