### 2019-08-22 574c766
Reverted deposition cloud to the grid-cloud from simpson 2015--works nicely with the weights, etc from Hopkins.  Spherical clouds... do not.
