+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -48,7 +48,7 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
     
     /* Compute size (in floats) of the current grid. */
-    float stretchFactor =1.15;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
+    float stretchFactor =1.;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
                                 // in units of dx. Since the cloud forms a sphere shell, stretchFactor > 1 is not recommended
     size = 1;
     for (int dim = 0; dim < GridRank; dim++)
@@ -124,20 +124,23 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     
     /* A DODECAHEDRON+ISOCAHEDRON */
 
-        int nCouple = 32;
+        int nCouple = 26;
         float A = stretchFactor*dx;
-
-        /* Dodec points followed by isoca points */
-        FLOAT CloudParticleVectorX [] = {1, 1, 1, 1, -1, -1, -1, -1, 0, 0, 0, 0,
-                                            iphi, iphi, -iphi, -iphi, phi, phi, -phi, -phi, 
-                                            0, 0, 0, 0, 1, 1, -1, -1, phi,-phi, phi,-phi};
-        FLOAT CloudParticleVectorY [] = {1,1,-1,-1, 1, 1, -1, -1, iphi, iphi, -iphi, -iphi,
-                                            phi, -phi, phi,-phi, 0, 0, 0, 0,1, 1, -1, -1, 
-                                            phi, -phi, -phi, phi, 0, 0, 0, 0};
-        FLOAT CloudParticleVectorZ []  = {1,-1, 1,-1, 1,-1, 1,-1, phi,-phi, phi,-phi, 
-                                            0, 0, 0, 0, iphi, -iphi, iphi, -iphi, 
-                                            phi, -phi, -phi, phi, 0, 0, 0, 0, 1, 1, -1, -1};
-                            
+        float cloudSize=1.*dx;
+        /* Points from HEALPix algorithm; 48 equally dist. points*/
+        FLOAT CloudParticleVectorX [] = {-1, -1, -1, -1, -1, -1, -1, -1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1};
+                                            // {1, 1, 1, 1, -1, -1, -1, -1, 0, 0, 0, 0,
+                                            // iphi, iphi, -iphi, -iphi, phi, phi, -phi, -phi, 
+                                            // 0, 0, 0, 0, 1, 1, -1, -1, phi,-phi, phi,-phi};
+        FLOAT CloudParticleVectorY [] = {1,1,1,0,0,-1,-1,-1,0,1,1,1,0,0,-1,-1,-1,1,1,1,0,0,-1,-1,-1,0};
+                                            // {1,1,-1,-1, 1, 1, -1, -1, iphi, iphi, -iphi, -iphi,
+                                            // phi, -phi, phi,-phi, 0, 0, 0, 0,1, 1, -1, -1, 
+                                            // phi, -phi, -phi, phi, 0, 0, 0, 0};
+        FLOAT CloudParticleVectorZ []  ={1,0,-1,1,-1,1,0,-1,0,1,0,-1,1,-1,1,0,-1,1,0,-1,1,-1,1,0,-1,0};
+                                            // {1,-1, 1,-1, 1,-1, 1,-1, phi,-phi, phi,-phi, 
+                                            // 0, 0, 0, 0, iphi, -iphi, iphi, -iphi, 
+                                            // phi, -phi, -phi, phi, 0, 0, 0, 0, 1, 1, -1, -1};
[…]
