### 2019-08-24 40f144c
Adding ThermalInstability problem type
