+++ src/enzo/ThermalInstabilityInitialize.C
@@ -1,0 +1,186 @@
+/***********************************************************************
+/
+/   Thermal Instability Test Problem
+/
+/   written by: Cameron Hummels, Iryna Butsky
+/   date:       March 2018
+/
+/   PURPOSE: Investigate how thermal instability operates when different
+/            physics are present
+/
+/   RETURNS: SUCCESS or FAIL
+/
+************************************************************************/
+ 
+#include <string.h>
+#include <stdio.h>
+#include <math.h>
+#include "ErrorExceptions.h"
+#include "macros_and_parameters.h"
+#include "typedefs.h"
+#include "global_data.h"
+#include "Fluxes.h"
+#include "GridList.h"
+#include "ExternalBoundary.h"
+#include "Grid.h"
+#include "Hierarchy.h"
+#include "TopGridData.h"
+
+void AddLevel(LevelHierarchyEntry *Array[], HierarchyEntry *Grid, int level);
+int RebuildHierarchy(TopGridData *MetaData,
+		     LevelHierarchyEntry *LevelArray[], int level);
+ 
+int ThermalInstabilityInitialize(FILE *fptr, FILE *Outfptr, HierarchyEntry &TopGrid,
+          TopGridData &MetaData)
+{
+   if(debug){
+      printf("Entering ThermalInstabilityInitialize\n");
+      fflush(stdout);
+      }
+
+   // Make sure that we are working in 3D
+   if (MetaData.TopGridRank != 3) {
+      ENZO_VFAIL("Cannot do ThermalInstability in %"ISYM" dimension(s)\n", MetaData.TopGridRank)
+      }
+
+   // Field Names
+   char *DensName = "Density";
+   char *TEName   = "TotalEnergy";
+   char *GEName   = "GasEnergy";
+   char *Vel1Name = "x-velocity";
+   char *Vel2Name = "y-velocity";
+   char *Vel3Name = "z-velocity";
+   char *BxName   = "Bx";
+   char *ByName = "By";
+   char *BzName = "Bz";
+   char *PhiName = "Phi";
+   char *CRName = "CREnergyDensity";
+   char *ElectronName = "Electron_Density";
+   char *HIName = "HI_Density";
+   char *HIIName = "HII_Density";
+   char *HeIName = "HeI_Density";
+   char *HeIIName = "HeII_Density";
+   char *HeIIIName = "HeIII_Density";
+   char *HMName = "HM_Density";
+   char *H2IName = "H2I_Density";
+   char *H2IIName = "H2II_Density";
+   char *DIName = "DI_Density";
+   char *DIIName = "DII_Density";
+   char *HDIName = "HDI_Density";
+   char *MetalName = "Metal_Density";
+
+   /* parameter declarations */
+ 
+   FLOAT LeftEdge[MAX_DIMENSION], RightEdge[MAX_DIMENSION];
+
+   // Local variable declarations
+   char line[MAX_LINE_LENGTH];
+   int   dim, ret, level;
+
+   // Initialize parameters to default values
+   float TIMeanDensity = 1;
+   float TIDensityPerturbationAmplitude = 0.01;
+   float TIMeanTemperature = 1000000;
+
+   // Read problem specific parameters. 
+   while (fgets(line, MAX_LINE_LENGTH, fptr) != NULL) {
+      ret = 0;
+
+      ret += sscanf(line, "TIMeanDensity = %"FSYM, &TIMeanDensity);
+      ret += sscanf(line, "TIDensityPerturbationAmplitude = %"FSYM, &TIDensityPerturbationAmplitude);
+      ret += sscanf(line, "TIMeanTemperature = %"FSYM, &TIMeanTemperature);
[…]
