+++ src/enzo/InitializeNew.C
@@ -155,6 +155,8 @@ int ShearingBox2DInitialize(FILE *fptr, FILE *Outfptr, HierarchyEntry &TopGrid,
                         TopGridData &MetaData);
 int ShearingBoxStratifiedInitialize(FILE *fptr, FILE *Outfptr, HierarchyEntry &TopGrid,
                         TopGridData &MetaData);
+int ThermalInstabilityInitialize(FILE *fptr, FILE *Outfptr, HierarchyEntry &TopGrid,
+                                 TopGridData &MetaData);
 #ifdef TRANSFER
 int PhotonTestInitialize(FILE *fptr, FILE *Outfptr, 
 			 HierarchyEntry &TopGrid, TopGridData &MetaData,
@@ -728,6 +730,9 @@ int InitializeNew(char *filename, HierarchyEntry &TopGrid,
   if ((ProblemType == 414) || (ProblemType == 415))
     ret = CosmoIonizationInitialize(fptr, Outfptr, TopGrid, MetaData, 0);
 
+  // 417 Thermal Instability initializer
+  if (ProblemType == 417)
+    ret = ThermalInstabilityInitialize(fptr, Outfptr, TopGrid, MetaData);
 
   // 450-452) Free-streaming radiation tests
   if ((ProblemType == 450) || (ProblemType == 451) || (ProblemType == 452))
