+++ src/enzo/Grid_ThermalInstabilityInitializeGrid.C
@@ -1,0 +1,352 @@
+/***********************************************************************
+/
+/   GRID CLASS (INITIALIZE THE GRID FOR THE THERMAL INSTABILITY TEST)
+/
+/   written by: Iryna Butsky and Cameron Hummels
+/   date:         June 2018
+/   modified1:   
+/
+/   PURPOSE: Sets up the grid for the ThermalInstability problem type.
+/            Your run directory must include the file 'white_noise.in', 
+/            which can be generated using the file 'white_noise_generator.py', 
+/            which should be in run/Hydro/Hydro-3D/ThermalInstability
+/
+/   RETURNS: FAIL or SUCCESS
+/
+************************************************************************/
+ 
+#include <stdio.h>
+#include <stdlib.h>
+#include <math.h>
+#include "ErrorExceptions.h"
+#include "macros_and_parameters.h"
+#include "typedefs.h"
+#include "global_data.h"
+#include "Fluxes.h"
+#include "GridList.h"
+#include "ExternalBoundary.h"
+#include "Grid.h"
+
+int GetUnits(float *DensityUnits, float *LengthUnits,
+	     float *TemperatureUnits, float *TimeUnits,
+	     float *VelocityUnits, double *MassUnits, FLOAT Time);
+
+int grid::ThermalInstabilityInitializeGrid(float TIMeanDensity,
+                                       float TIDensityPerturbationAmplitude,
+                                       float TIMeanTemperature)
+{
+   if (ProcessorNumber != MyProcessorNumber)
+      return SUCCESS;
+
+   if(debug){
+      printf("Entering ThermalInstabilityInitializeGrid\n");
+      fflush(stdout);
+      }
+ 
+   // Figure out grid quantities and how to access fields 
+   int size = 1;
+
+   for (int dim = 0; dim < GridRank; dim++)
+      size *= GridDimension[dim];
+
+   int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num, B1Num, B2Num, B3Num, PhiNum;
+   int DeNum, HINum, HIINum, HeINum, HeIINum, HeIIINum, HMNum, H2INum, H2IINum,
+      DINum, DIINum, HDINum, MetalNum;
+
+   if (this->IdentifyPhysicalQuantities(DensNum, GENum, Vel1Num, Vel2Num,
+					Vel3Num, TENum, B1Num, B2Num, B3Num, PhiNum) == FAIL) {
+      ENZO_FAIL("Error in IdentifyPhysicalQuantities.\n");
+      }
+
+   int MetallicityField = FALSE;
+
+   if (MultiSpecies) {
+      if (IdentifySpeciesFields(DeNum, HINum, HIINum, HeINum, HeIINum, HeIIINum,
+                      HMNum, H2INum, H2IINum, DINum, DIINum, HDINum) == FAIL) {
+         ENZO_FAIL("Error in grid->IdentifySpeciesFields.");
+         }
+      }
+
+   if ((MetalNum = FindField(Metallicity, FieldType, NumberOfBaryonFields))
+         != -1)
+      MetallicityField = TRUE;
+   else
+      MetalNum = 0;
+
+   // Get the units
+   float DensityUnits, LengthUnits, TemperatureUnits = 1, TimeUnits, VelocityUnits; 
+   double MassUnits;
+   
+   GetUnits(&DensityUnits, &LengthUnits, &TemperatureUnits,
+	    &TimeUnits, &VelocityUnits, &MassUnits, Time);
+
+   printf("Getting ready to set up the thermal instability box.\n");
+
+   // Set up the white noise field.
+   int pert_size_x;
+   int pert_size_y;
+   int pert_size_z;
+
[…]
