### 2019-08-26 #1636 @danielskatz
That's ok, but JOSS reviews typically are more interactive.  Maybe with a complex team like this, it makes sense to wait.
