### 2019-08-27 f0a77bf
adding cosmic ray second fluid; tested SOD shocktube
