+++ src/enzo/hydro_rk/Grid_UpdateMHDPrim.C
@@ -37,10 +37,15 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
   }
 
   int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num, 
-    B1Num, B2Num, B3Num, PhiNum;
+    B1Num, B2Num, B3Num, PhiNum, CRNum;
 
   this->IdentifyPhysicalQuantities(DensNum, GENum, Vel1Num, Vel2Num, 
 				   Vel3Num, TENum, B1Num, B2Num, B3Num, PhiNum);
+
+  if (CRModel) {
+    if ((CRNum = FindField(CRDensity, FieldType, NumberOfBaryonFields)) < 0)
+      ENZO_FAIL("Cannot Find Cosmic Rays");
+  }
 
   int i, j, k, n, n_dU, dim, igrid, field, size, activesize;
   for (dim = 0, size = 1; dim < GridRank; dim++) {
@@ -133,7 +138,7 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
     rho, vx, vy, vz, e, etot, Tau, eint, p, v2,
     D_new, S1_new, S2_new, S3_new, Tau_new, h, cs, dpdrho, dpde, Eint_new,
     Bx_old, By_old, Bz_old, Bx, By, Bz, Bx_new, By_new, Bz_new,
-    Phi_old, Phi, Phi_new, B2;
+    Phi_old, Phi, Phi_new, B2, ecr_old, ecr, ECR_new;
 
   float rhou, lenu, tempu, tu, velu;
   GetUnits(&rhou, &lenu, &tempu, &tu, &velu, Time);
@@ -162,6 +167,9 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
 	By_old   = OldBaryonField[B2Num][igrid];
 	Bz_old   = OldBaryonField[B3Num][igrid];
 	Phi_old  = OldBaryonField[PhiNum][igrid];
+	if (CRModel){
+          ecr_old = OldBaryonField[CRNum][igrid];
+	}
 
 	rho  = BaryonField[DensNum][igrid];
 	vx   = BaryonField[Vel1Num][igrid];
@@ -176,6 +184,9 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
 	By   = BaryonField[B2Num][igrid];
 	Bz   = BaryonField[B3Num][igrid];
 	Phi  = BaryonField[PhiNum ][igrid];
+        if (CRModel){
+          ecr = BaryonField[CRNum][igrid];
+	}
 
 	D_new   = c1*rho_old + (1.0-c1)*rho + c2*dU[iD][n];
 	S1_new  = c1*rho_old*vx_old + (1.0-c1)*rho*vx + c2*dU[iS1][n];
@@ -189,6 +200,8 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
 	if (DualEnergyFormalism) {
 	  Eint_new = c1*rho_old*eint_old + (1.0-c1)*rho*eint + c2*dU[iEint][n];
 	}
+	if (CRModel)
+          ECR_new = c1*ecr_old + (1.0-c1)*ecr + c2*dU[iCR][n];
 
 	if (D_new < 0 || isnan(D_new)) {
 
@@ -206,7 +219,12 @@ int grid::UpdateMHDPrim(float **dU, float c1, float c2)
 	vy = S2_new/D_new;
 	vz = S3_new/D_new;
 	etot = Tau_new/D_new;
-
+        if (CRModel){
[…]
