+++ src/enzo/global_data.h
@@ -424,16 +424,23 @@ EXTERN int MultiMetals;
  * 1: On, (two fluid model)
  */
 EXTERN int CRModel;
+
 /* Cosmic Ray Diffusion
  * 0: Off - default
  * 1: On, CRkappa is constant across grid
+ * 2: On, anisotropic diffusion with constant CRkappa
  */
 EXTERN int CRDiffusion;
+
 /* Cosmic Ray Feedback
  *    0.0 -- No CR feedback
  *    1.0 -- All feedback into CR field
  */
 EXTERN float CRFeedback;
+
+EXTERN int CRHeating; // 0 is off, 1 is on
+EXTERN int CRStreaming; // 0 is off, 1 is on 
+EXTERN float CRStreamingFactor;
 EXTERN float CRkappa;
 EXTERN float CRCourantSafetyNumber;
 EXTERN float CRdensFloor;
@@ -835,6 +842,7 @@ EXTERN int iS1;
 EXTERN int iS2;
 EXTERN int iS3;
 EXTERN int iEint;
+EXTERN int iCR;
 EXTERN float SmallRho;
 EXTERN float SmallP;
 EXTERN float SmallEint;
