+++ src/enzo/ReadParameterFile.C
@@ -634,15 +634,18 @@ int ReadParameterFile(FILE *fptr, TopGridData &MetaData, float *Initialdt)
       ret++;
     }
 
-    ret += sscanf(line, "CRModel = %"ISYM, &CRModel); 
+    ret += sscanf(line, "CRModel = %"ISYM, &CRModel);
     ret += sscanf(line, "CRDiffusion = %"ISYM, &CRDiffusion);
+    ret += sscanf(line, "CRHeating = %"ISYM, &CRHeating);
+    ret += sscanf(line, "CRStreaming = %"ISYM, &CRStreaming);
+    ret += sscanf(line, "CRStreamingFactor = %"FSYM, &CRStreamingFactor);
     ret += sscanf(line, "CRkappa = %"FSYM, &CRkappa);
     ret += sscanf(line, "CRCourantSafetyNumber = %"FSYM, &CRCourantSafetyNumber);
     ret += sscanf(line, "CRFeedback = %"FSYM, &CRFeedback);
     ret += sscanf(line, "CRdensFloor = %"FSYM, &CRdensFloor);
     ret += sscanf(line, "CRmaxSoundSpeed = %"FSYM, &CRmaxSoundSpeed);
     ret += sscanf(line, "CRgamma = %"FSYM, &CRgamma);
-    ret += sscanf(line, "CosmologySimulationUniformCR = %"FSYM, &CosmologySimulationUniformCR); // FIXME
+    ret += sscanf(line, "CosmologySimulationUniformCR = %"FSYM, &CosmologySimulationUniformCR); // FIXME  
 
     ret += sscanf(line, "ShockMethod = %"ISYM, &ShockMethod);
     ret += sscanf(line, "ShockTemperatureFloor = %"FSYM, &ShockTemperatureFloor);
@@ -1935,6 +1938,12 @@ int ReadParameterFile(FILE *fptr, TopGridData &MetaData, float *Initialdt)
     iEint = 5;
   }
 
+  if (( CRModel ) && ( HydroMethod == 4 )){
+    // to be used with DualEnergyFormalism                                                                                         
+    NEQ_MHD = 11;
+    iCR = 10;
+  }
+
   // Determine color fields (NColor) later inside a grid object.
   // Don't include free electron field
   switch (MultiSpecies) {
@@ -2173,6 +2182,10 @@ int ReadParameterFile(FILE *fptr, TopGridData &MetaData, float *Initialdt)
   /* Cosmic ray diffusion should be off if Cosmic rays are off */
   if(CRDiffusion > 0 && CRModel == 0){
     ENZO_FAIL("CRDiffusion can only be used if CRModel is turned on!!\n");
+  }
+
+  if (CRModel == 1  &&  HydroMethod == 4 && DualEnergyFormalism == 0){
+    ENZO_FAIL("CR physics can only be used with HydroMethod = 4 if DualEnergyFormalism is turned on!\n");
   }
 
   if (debug) printf("Initialdt in ReadParameterFile = %e\n", *Initialdt);

[…]
