### 2019-08-29 enzo-project/enzo-dev#109 @rtfisher
[JOSS Review] Quickstart documentation 
The quickstart documentation on the collapse problem would benefit from graphic plots showing the yt plot output in section 2.1.5, so that users can confirm their results visually.

JOSS Review: openjournals/joss-reviews#1636
