### 2019-08-29 #1636 @rtfisher
Ok, I've completed my initial review. It looks like I also identified some of the same issues as @zingale, though our final assessments of the JOSS requirements differ slightly.
