### 2019-08-30 ec89faf
Added documentation for Enzo's hybrid mode
