### 2019-09-06 1b12cec
Added APM implementation from my hg repo + Greg’s fork. Results from the paper are reproduced. Additional: * changes to make APM tests completely isolated * tests added to the test suite * updated documentation + authors list * updated .gitignore * show log in case of error in circleci
