+++ src/enzo/ComputeAccelerationFieldAPMGravity.C
@@ -1,0 +1,265 @@
+/***********************************************************************
+/
+/  COMPUTE ACCELERATION FIELD FUNCTION
+/
+/  written by: Greg Bryan
+/  date:       March, 1995
+/  modified1:
+/
+/  PURPOSE:
+/
+************************************************************************/
+
+#ifdef USE_MPI
+#include <mpi.h>
+#endif /* USE_MPI */
+
+#include <stdio.h>
+#include <stdlib.h>
+#include "ErrorExceptions.h"
+#include "EnzoTiming.h"
+#include "performance.h"
+#include "macros_and_parameters.h"
+#include "typedefs.h"
+#include "global_data.h"
+#include "Fluxes.h"
+#include "GridList.h"
+#include "ExternalBoundary.h"
+#include "Grid.h"
+#include "Hierarchy.h"
+#include "LevelHierarchy.h"
+#include "TopGridData.h"
+#include "communication.h"
+#include "CommunicationUtilities.h"
+
+/* function prototypes */
+
+int CommunicationReceiveHandler(fluxes **SubgridFluxesEstimate[] = NULL,
+                                int NumberOfSubgrids[] = NULL,
+                                int FluxFlag = FALSE,
+                                TopGridData* MetaData = NULL);
+
+/* EvolveHierarchy function */
+
+int ComputeAccelerationFieldAPMGravity(HierarchyEntry *Grid, TopGridData *MetaData,
+			     LevelHierarchyEntry *LevelArray[], int level,
+			     ExternalBoundary *Exterior = NULL)
+{
+
+  /* declarations */
+
+  int dim;
+  grid *CurrentGrid = Grid->GridData;
+
+  /* Compute the refinement factor. */
+
+  int RefinementFactor = 1, Factors[MAX_DIMENSION];
+  if (Grid->ParentGrid != NULL) {
+    Grid->ParentGrid->GridData->ComputeRefinementFactors(CurrentGrid, Factors);
+    for (dim = 0; dim < MAX_DIMENSION; dim++) {
+      if (Factors[dim] != 1) 
+	RefinementFactor = Factors[dim];
+    }
+  }
+
+  /* Compute the acceleration (and potential) on the grid. */
+  if (level == 0) {
+
+    /* Here we calculate the acceleration field for PPM 
+       which is assumed to be face-centered.
+       If Zeus is used, staggering will be done in ZeusSource.C */
+    int DiffType = DIFFERENCE_TYPE_NORMAL;
+    if (CurrentGrid->ComputeAccelerationField(DiffType, level) == FAIL)
+      ENZO_FAIL("Error in grid->ComputeAccelerationField.C\n");
+    
+  } else {
+    if (CurrentGrid->ComputeAccelerationFieldAPM(RefinementFactor) == FAIL)
+      ENZO_FAIL("Error in grid->ComputeAccelerationFieldAPM.\n");
+  
+    /* For the TestSelfForce we may want to output the force components separately */
+    if (APMAddParentContribution) {
+
+      
+      /* a) Add parental acceleration. */
+      /* Three-pass structure based on SetBoundaryConditions.C */
+      LCAPERF_START("APMAddParentAcceleration");
+      TIMER_START("APMAddParentAcceleration");
+
+#ifdef FORCE_MSG_PROGRESS
+      CommunicationBarrier();
+#endif
+      
+      /* -------------- FIRST PASS ----------------- */
+      /* Here, we just generate the calls to generate the receive buffers,
+         without actually doing anything. */
+      
+      CommunicationDirection = COMMUNICATION_POST_RECEIVE;
[…]
