+++ src/enzo/TestGravitySphereAPMInitialize.C
@@ -1,0 +1,222 @@
+/***********************************************************************
+/
+/  INITIALIZE A GRAVITY TEST
+/
+/  written by: Greg Bryan
+/  date:       September, 1995
+/  modified1: Jean-Claude Passy, June 2013
+/
+/  PURPOSE:
+/    Set up a spherical mass distribution to test the gravity solver.
+/
+/  UPDATE: Following tests performed in Passy & Bryan (2014).
+/          Possible distributions are:
+/           - uniform density (type 0)
+/           - isothermal spere (type 1)
+/           - Plummer sphere (type 2)
+/
+/  RETURNS: SUCCESS or FAIL
+/
+************************************************************************/
+
+// This routine intializes a new simulation based on the parameter file.
+//
+
+#include <string.h>
+#include <stdio.h>
+#include <math.h>
+#include "ErrorExceptions.h"
+#include "macros_and_parameters.h"
+#include "typedefs.h"
+#include "global_data.h"
+#include "Fluxes.h"
+#include "GridList.h"
+#include "ExternalBoundary.h"
+#include "Grid.h"
+#include "Hierarchy.h"
+#include "LevelHierarchy.h"
+#include "TopGridData.h"
+
+void AddLevel(LevelHierarchyEntry *Array[], HierarchyEntry *Grid, int level);
+int RebuildHierarchy(TopGridData *MetaData,
+                     LevelHierarchyEntry *LevelArray[], int level);
+
+int TestGravitySphereAPMInitialize(FILE *fptr, FILE *Outfptr,
+                                   HierarchyEntry &TopGrid, TopGridData &MetaData)
+{
+  char *DensName = "Density";
+  char *TEName = "TotalEnergy";
+  char *GEName = "GasEnergy";
+  char *Vel1Name = "x-velocity";
+  char *Vel2Name = "y-velocity";
+  char *Vel3Name = "z-velocity";
+  char *GPotName = "Grav_Potential";
+
+  /* declarations */
+
+  char line[MAX_LINE_LENGTH];
+  int dim, ret, level;
+
+  /* Error check. */
+
+  if (!SelfGravity)
+    fprintf(stderr, "TestGravitySphere: gravity is off!?!");
+
+  /* set default parameters */
+
+  float TestGravitySphereInteriorDensity = 1.0;         // density inside sphere
+  float TestGravitySphereExteriorDensity = tiny_number; // density outside sphere
+  float TestGravitySphereRadius = 0.1;
+  int TestGravitySphereType = 0; // uniform density
+  int TestGravitySphereUseBaryons = TRUE;
+  int TestGravitySphereRefineAtStart = FALSE;
+  FLOAT TestGravitySphereCenter[MAX_DIMENSION];
+  for (dim = 0; dim < MAX_DIMENSION; dim++)
+    TestGravitySphereCenter[dim] = 0.5 * (DomainLeftEdge[dim] + DomainRightEdge[dim]);
+
+  /* read input from file */
+
+  while (fgets(line, MAX_LINE_LENGTH, fptr) != NULL)
+  {
+
+    ret = 0;
+
+    /* read parameters */
+
+    ret += sscanf(line, "TestGravitySphereInteriorDensity = %" FSYM,
+                  &TestGravitySphereInteriorDensity);
+    ret += sscanf(line, "TestGravitySphereExteriorDensity = %" FSYM,
+                  &TestGravitySphereExteriorDensity);
+    ret += sscanf(line, "TestGravitySphereRadius = %" FSYM,
+                  &TestGravitySphereRadius);
+    ret += sscanf(line, "TestGravitySphereType = %" ISYM,
+                  &TestGravitySphereType);
+    ret += sscanf(line, "TestGravitySphereUseBaryons = %" ISYM,
+                  &TestGravitySphereUseBaryons);
[…]
