+++ run/GravitySolver/GravityTest/GravityTest__test_gravity.py
@@ -7,7 +7,7 @@ from yt.frontends.enzo.answer_testing_support import \
      requires_outputlog
 
 tolerance = ytcfg.getint("yt", "answer_testing_tolerance")
-     
+
 _pf_name = os.path.basename(os.path.dirname(__file__)) + ".enzo"
 _dir_name = os.path.dirname(__file__)
 
@@ -20,9 +20,9 @@ class TestGravityTest(AnswerTestingTest):
 
     def run(self):
         Data = np.loadtxt("TestGravityCheckResults.out")
-        radius = Data[:,0]
-        ForceRadialComputed = Data[:,2]
-        ForceRadialTrue = Data[:,3]
+        radius = Data[:,4]
+        ForceRadialComputed = Data[:,6]
+        ForceRadialTrue = Data[:,7]
 
         Error = (ForceRadialComputed-ForceRadialTrue)/ForceRadialTrue
         indices = np.where((radius > 1.0) & (radius < 8.0))

[…]
