### 2019-09-11 e888e4e
Fixing the quickstart JOSS issue with the Collapse Test
