### 2019-09-11 enzo-project/enzo-dev#98 @fearmayo
I'll take this on and roll into the same PR as #109 and #99
