### 2019-09-11 a11c438
A fix for issue #98, where yt references are updated in the quickstart guide
