+++ doc/manual/source/user_guide/bootcamp.rst
@@ -1,7 +1,7 @@
 Quickstart Guide to Enzo
 ========================
 
-This is an unofficial quickstart guide to using Enzo, starting from
+This is the quickstart guide to using Enzo, starting from
 checking out the Enzo and YT sources, through to building them, then
 to running an example problem and looking at its output.
 
@@ -10,7 +10,7 @@ useful information or useful tools:
  
 * `github.com/enzo-project/enzo-dev <http://github.com/enzo-project/enzo-dev>`_
 * `yt-project.org <http://yt-project.org>`_
-* `jacques.enzotools.org <http://jacques.enzotools.org>`_
+
 
 If you ever get stuck, please email `enzo-users
 <https://groups.google.com/forum/#!forum/enzo-users>`_. This is not
@@ -21,7 +21,15 @@ interlocking modules, unusual combinations of parameters may require
 some care, and the individual developers that have created those
 modules will be able to discuss them with you.
 
-
+Getting YT
+------------
+YT was initially developed to allow fast and easy analysis of Enzo output. Since it's
+original inception YT has grown to be a powerful tool capable of analysing
+many different astrophysical simulation code outputs and also non-astro
+simulation outputs too.
+To get your hands on YT (which you will need to complete this quickstart tutorial)
+then go to the `YT installation page <https://yt-project.org/doc/installing.html>`_
+and install YT locally.
 
 Getting Enzo
 ------------
