### 2019-09-11 enzo-project/enzo-dev#108 @bwoshea
@rtfisher  I want to push back on this.  **I don't think this is actually a problem.**  When I look at other JOSS papers (see, e.g., https://joss.theoj.org/papers/571753bc54c5d6dd36382c3d801de41d) the associated markdown file (https://github.com/sigsep/open-unmix-paper-joss/blob/master/paper.md) looks the same as what we have, although the PDF has references that come from the associated BibTex file (using the keys in the main text).  There are other papers that have a mixture of BibTex references and footnote-style references (to, e.g., URLS) - see https://joss.theoj.org/papers/9a87d84b2ed00ed82a6e297a4c34b3cf and the paper at https://github.com/WRF-CMake/wrf/blob/wrf-cmake/paper/paper.md for an example.  It's also worth noting that the Markdown paper that JOSS refers to in its author information (https://joss.readthedocs.io/en/latest/submitting.html) has the exact same outcome as ours w.r.t. the References section.

Anyway, would you take a look at the other papers I've cited, and at the JOSS author instructions, and confirm if I'm correct here?  Thanks!
