### 2019-09-11 enzo-project/enzo-dev#94 @bwoshea
This is addressed in PR #113  @zingale
