### 2019-09-12 f35dd3f
dealt with a variety of documentation issues
