### 2019-09-13 529d9e5
Add ubuntu machine file and update CI to use it
