+++ .circleci/config.yml
@@ -46,7 +46,7 @@ commands:
           cd $HOME/grackle
           ./configure
           cd src/clib
-          make machine-linux-gnu
+          make machine-ubuntu
           make
           make install
 
@@ -91,7 +91,7 @@ commands:
             git checkout << parameters.tag >>
             ./configure
             cd src/enzo
-            make machine-linux-gnu
+            make machine-ubuntu
             make load-config-test-suite
             make clean
             make -j 4
@@ -107,7 +107,7 @@ commands:
           source $HOME/venv/bin/activate
           ./configure
           cd src/enzo
-          make machine-linux-gnu
+          make machine-ubuntu
           # test various baryon/particle precisions
           for prec in 32 64
           do for part in 32 64 128
