### 2019-09-13 b274754
Revert grackle CI machine config
