### 2019-09-13 enzo-project/enzo-dev#96 @bwoshea
@zingale  This is addressed in PR #115
