### 2019-09-13 616710c
Update doc/manual/source/index.rst
Co-Authored-By: Greg Bryan <gbryan@astro.columbia.edu>
