### 2019-09-13 c59599d
Update doc/manual/source/physics/adaptive_mesh.rst
Co-Authored-By: Greg Bryan <gbryan@astro.columbia.edu>
