### 2019-09-13 89680c2
Update doc/manual/source/physics/adaptive_mesh.rst
Co-Authored-By: Britton Smith <brittonsmith@gmail.com>
