### 2019-09-13 #1636 @bwoshea
@zingale @rtfisher just to give you a heads-up, we've addressed all of the issues you have raised but one, and that should be done in the next day or so.  We've left all off the PRs open so that you can see what has been changed, and all PRs and issues are cross-referenced.  We're ready for your feedback whenever you are available.
