### 2019-09-13 97fbe68
adding test problem descriptions to docs
