+++ doc/manual/source/testproblems/index.rst
@@ -1,0 +1,766 @@
+.. _parameters:
+
+Enzo Test Problems
+===================
+
+The following is a list of the test problems that are available in the 
+run directory found in the Enzo repository. These problems are also the
+same set of problems used in the test suite. They are listed in broad
+categories following the directory structure.
+
+This list however is not a complete list of all the ProblemTypes that 
+are available in Enzo.
+
+The test problem specific  parameters can be found in the Parameter List.
+
+.. toctree::
+   :maxdepth: 2
+
+
+* `Cooling`_ 
+
+   * `CoolingTest_Cloudy`_
+
+   * `CoolingTest_Grackle`_ 
+
+   * `CoolingTest_JHW`_  
+
+   * `OneZoneFreefallTest`_
+
+
+* `Cosmology`_
+
+   * `AdiabaticExpansion`_
+
+   * `AMRZeldovichPancake`_  
+   
+   * `AMRZeldovichPancake_Streaming`_
+
+   * `MHDAdiabaticExpansion_CT`_       
+   
+   * `MHDAdiabaticExpansion_Dedner`_  
+   
+   * `MHDZeldovichPancake`_           
+
+   * `MHDZeldovichPancake_2_CT`_      
+
+   * `MHDZeldovichPancake_2_Dedner`_
+
+   * `SphericalInfall`_
+
+   * `ZeldovichPancake`_
+
+
+* `Cosmology Simulation`_
+
+   * `amr_cosmology`_
+
+   * `amr_nested_cosmology`_
+
+   * `dm_only`_
+
+   * `ReionizationHydro`_
+
+   * `ReionizationRadHydro`_
+
+
+* `DrivenTurbulence3D`_
+
+* `FLD`_
+
+* `FuzzyDarkMatter`_
+
+* `Gravity Solver`_
+
+   * `BinaryCollapse`_  
+
+   * `BinaryCollapseMHDCT`_  
+
+   * `GravityStripTest`_  
+
+   * `GravityTest`_  
+
+   * `GravityTestSphere`_  
+
+   * `MaximumGravityRefinementTest`_  
+
+   * `TestOrbit`_  
+
+   * `TestOrbitMRP`_
+
+* `Hydro/Hydro-1D`_
+
+   * `FreeExpansion`_
+
+   * `InteractingBlastWaves`_
+
+   * `PressurelessCollapse`_
+
+   * `ShockInABox`_
+
+   * `SodShockTube`_
+
+   * `Toro-1-ShockTube`_
+
+   * `Toro-1-ShockTubeAMR`_
+
+   * `Toro-2-ShockTube`_
+
+   * `Toro-2-ShockTubeAMR`_
+
+   * `Toro-3-ShockTube`_
+
+   * `Toro-3-ShockTubeAMR`_
+
+   * `Toro-4-ShockTube`_
+
+   * `Toro-4-ShockTubeAMR`_
+
+   * `Toro-5-ShockTube`_
+
+   * `Toro-5-ShockTubeAMR`_
+
+   * `Toro-6-ShockTube`_
+
+   * `Toro-7-ShockTube`_
+
+   * `WavePool`_
+
+* `Hydro/Hydro-2D`_
+
+   * `AMRShockPool2D`_
+
+   * `Athena-RayleighTaylor`_
+
+   * `DoubleMachReflection`_
+
+   * `FreeExpansionAMR`_
+
+   * `HDMHD2DCheckOddEvenCouplingOfRiemannSolver`_
+
+   * `Implosion`_
+
+   * `ImplosionAMR`_
+
+   * `KelvinHelmholtz`_
+
+   * `KelvinHelmholtzAMR`_
+
[…]
