+++ doc/manual/source/testproblems/index.rst
@@ -1,4 +1,4 @@
-.. _parameters:
+.. _test_problems:
 
 Enzo Test Problems
 ===================
@@ -192,7 +192,7 @@ The test problem specific  parameters can be found in the Parameter List.
 
    * `StripTest`_
 
-* `MHD/MHD-1D`_
+* `MHD/1D`_
 
    * `BrioWu-MHD-1D`_
 
@@ -204,7 +204,7 @@ The test problem specific  parameters can be found in the Parameter List.
 
    * `MHD_Metal_Advection_Dedner`_
 
-* `MHD/MHD-2D`
+* `MHD/2D`
 
    * `LoopAdvection_CT`_
 
@@ -226,7 +226,7 @@ The test problem specific  parameters can be found in the Parameter List.
 
    * `Wengen2-CollidingFlow`_
 
-* `MHD/MHD-3D`
+* `MHD/3D`
 
    * `ShearingBox`_
 
@@ -471,11 +471,37 @@ Modified to add DM particles and a bulk gas velocity
 MHDAdiabaticExpansion_CT
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
+Adiabatic expansion test for MHD, using Athena CT.
+
+This test differs from the PPM Adiabatic Expansion by the increased initial
+temperature.  The extremely low floor on the PPM version makes it a poor test,
+since the thermal expansion is dominated by the temperature floor, rather than
+physical integration. 
+
+This test is not entirely uniform for two reasons.  
+First is self gravity (actually off in this version) which causes issues at the
+corners of the domain as well as subgrid boundary.
+The second is the time-interpolation in the boundary for the subgrid, which
+causes slight acceleration due to slightly different expansion of the fluid in
+the boundary of the subgrid relative to the "ideal" solution.
+
 .. _MHDAdiabaticExpansion_Dedner:
 
 MHDAdiabaticExpansion_Dedner
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Adiabatic expansion test for MHD, using Dedner.
 
+This test differs from the PPM Adiabatic Expansion by the increased initial
+temperature.  The extremely low floor on the PPM version makes it a poor test,
+since the thermal expansion is dominated by the temperature floor, rather than
+physical integration. 
+
+This test is not entirely uniform for two reasons.  
+First is self gravity (actually off in this version) which causes issues at the
+corners of the domain as well as subgrid boundary.
+The second is the time-interpolation in the boundary for the subgrid, which
+causes slight acceleration due to slightly different expansion of the fluid in
+the boundary of the subgrid relative to the "ideal" solution.
 
 .. _MHDZeldovichPancake:
 
@@ -487,12 +513,28 @@ MHDZeldovichPancake
 
 MHDZeldovichPancake_2_CT
 ^^^^^^^^^^^^^^^^^^^^^^^^
+This is another iteration of Zel'Dovich pancake.  This is tuned to almost 
+reproduce the result from Collins et al 2010, as well as the Dedner run in 
+run/Cosmology/MHDZeldovichPancake_2_Dedner
 
+Slight differences with the method paper exist due to the uniform initial
+distribution in the method paper.  
+
+Neither this nor the Dedner version use Dual Energy Formalism, in order to match
+the temperature field as well as possible.
 .. _MHDZeldovichPancake_2_Dedner:
 
 MHDZeldovichPancake_2_Dedner
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
[…]
