### 2019-09-15 enzo-project/enzo-dev#106 @bwoshea
@zingale this is now addressed in PR #118 .  One issue with documenting every problem in the run/ directory is some subdirectories have many variants of the same problem (e.g., shock tubes, I-front propagation, etc.), and the run/ directory gets fairly regularly updated.  It may make more sense to have documentation of each of the top-level subdirectories within the run/ directory, as well as highlighting a few of the standard/more interesting test problems, and then ensure that the parameter file for each test problem contains an explanation of what it's testing.  That would help with documentation in some ways - i.e., when people put new tests in, all they have to do is make sure the parameter file has comments rather than update both the run/ directory and the documentation.  What do you think?
