+++ doc/manual/source/testproblems/index.rst
@@ -11,7 +11,7 @@ categories following the directory structure.
 This list however is not a complete list of all the ProblemTypes that 
 are available in Enzo.
 
-The test problem specific  parameters can be found in the Parameter List.
+The test problem specific  parameters can be found in the :ref:`parameters`.
 
 .. toctree::
    :maxdepth: 2
@@ -154,7 +154,7 @@ The test problem specific  parameters can be found in the Parameter List.
 
    * `RampedKelvinHelmholtz2D`_
 
-   * `SedovBlasti`_
+   * `SedovBlast`_
 
    * `SedovBlastAMR`_
 
@@ -308,7 +308,7 @@ The test problem specific  parameters can be found in the Parameter List.
 * `StarParticle`_
 
 
-.. _Cooling:
+.. _CoolingProblems:
 
 Cooling
 ~~~~~~~
@@ -326,14 +326,12 @@ type did not exist in Enzo 1.5, so there is no comparison.
 The cooling tests will run in a few minutes on a single processor.
 
 The three parameter files are:
-CoolingTest_Cloudy.enzo - uses Cloudy cooling along with the 
-			       	      	    MultiSpecies = 1 chemistry.  The input data 
-					    provided is a three dimensional table that 
-					    varies in density, metallicity, and temperature.
+CoolingTest_Cloudy.enzo - uses Cloudy cooling along with the MultiSpecies = 1 chemistry.  
+The input data provided is a three dimensional table that varies in density, metallicity, 
+and temperature.
 
 Cooling data files:
-primordial_cie.dat - CIE cooling rates for atomic H and He taken from 
-		     	 	 Black (1981).
+primordial_cie.dat - CIE cooling rates for atomic H and He taken from Black (1981).
 solar_2008_3D_metals.h5 - input data for Cloudy cooling.
 
 The script plot.py will plot cooling rates from the cooling test 
@@ -354,11 +352,11 @@ The cooling tests will run in a few minutes on a single processor.
 
 The three parameter files are:
 CoolingTest_Grackle.enzo - uses Grackle cooling along with the 
-                           non-equilibrium atomic H/He chemistry.
+non-equilibrium atomic H/He chemistry.
 
 Cooling data files:
 primordial_cie.dat - CIE cooling rates for atomic H and He taken from 
-		     	 	 Black (1981).
+Black (1981).
 CloudyData_UVB=HM2012.h5 - input data for Grackle cooling.
 
 The script plot.py will plot cooling rates from the cooling test 
@@ -380,13 +378,13 @@ The cooling tests will run in a few minutes on a single processor.
 
 The three parameter files are:
 CoolingTest_JHW.enzo - uses John Wise's metal cooling along with 
-		       	    	 	the MultiSpecies = 1 chemitry.
+the MultiSpecies = 1 chemitry.
 
 Cooling data files:
[…]
