+++ doc/manual/source/user_guide/RunTestProblem.rst
@@ -24,7 +24,8 @@ simulation.
 
 The test problem parameter files are inside the ``run`` subdirectory.
 Please see :ref:`EnzoTestSuite` for a full list of test
-problems. The files that end in .enzo are the Enzo parameter files,
+problems. Descriptions for the test problems can be found in 
+:ref:`test_problems`. The files that end in .enzo are the Enzo parameter files,
 and .inits are inits parameter files. inits files are only used for
 cosmology simulations, and you can see an example of how to run
 that in :ref:`RunCosmologySimulation`. Let's try a
