### 2019-09-16 enzo-project/enzo-dev#97 @zingale
this change looks good
