### 2019-09-16 enzo-project/enzo-dev#100 @zingale
looks good
