### 2019-09-16 enzo-project/enzo-dev#106 @zingale
yes, I think that's a good idea.
