### 2019-09-16 #1636 @zingale
It looks like all of my issues have been addressed via PRs.  Once those PRs are merged, I am happy with accepting.
