### 2019-09-16 #1636 @rtfisher
I am fine with accepting after merger as well.
