### 2019-09-16 #1636 @bwoshea
@zingale @rtfisher thanks!

@rtfisher to clarify, there are two open issues that you raised (https://github.com/enzo-project/enzo-dev/issues/108 and https://github.com/enzo-project/enzo-dev/issues/109 ) - would you mind making a quick note saying you're OK with what we've done for those?
