### 2019-09-17 enzo-project/enzo-dev#108 @rtfisher
Ok, sounds good. I concur.
