### 2019-09-17 #1636 @rtfisher
Thank you, @bwoshea. Sorry for the delayed response -- I was serving on a panel today.

I concur that the issues are now resolved, and have closed both.
