### 2019-09-18 76fe2a3
Added seeding field for PIII stars; separately track SNII metal, SNIa metals, and P3 metals in different fields.
