+++ src/enzo/MechStars_transformComovingWithStar.C
@@ -6,10 +6,12 @@
 
 
 int transformComovingWithStar(float* Density, float* Metals, 
+        float* MetalsSNII, float* MetalsSNIA,
         float* Vel1, float* Vel2, float* Vel3, 
         float up, float vp, float wp,
         int sizeX, int sizeY, int sizeZ, int direction){
-    /* transform velocities to momenta or back and make them comoving with the star particle */
+    /* transform velocities to momenta or back and make them comoving with the star particle 
+        Transform metallicity fields to metal density fields*/
     if (direction > 0){
 
         /* To comoving with star */
@@ -23,13 +25,15 @@ int transformComovingWithStar(float* Density, float* Metals,
                     Vel2[ind] = (preV-vp)*Density[ind];
                     preV = Vel3[ind];
                     Vel3[ind] = (preV-wp)*Density[ind];
+                    MetalsSNII[ind] = Metals[ind]*Density[ind];
+                    MetalsSNIA[ind] = MetalsSNIA[ind]*Density[ind];
                 }
             }
         }
     }
     if (direction < 0){
 
-        /* back to "lab" frame */
+        /* back to "lab" frame, metal densities back to metallicities */
         for (int k = 0; k < sizeZ; k++){
             for (int j = 0; j<sizeY; j++){
                 for (int i = 0; i<sizeX; i++){
@@ -37,6 +41,8 @@ int transformComovingWithStar(float* Density, float* Metals,
                     Vel1[ind] = Vel1[ind]/Density[ind]+up;
                     Vel2[ind] = Vel2[ind]/Density[ind]+vp;
                     Vel3[ind] = Vel3[ind]/Density[ind]+wp;
+                    MetalsSNII[ind] = Metals[ind]/Density[ind];
+                    MetalsSNIA[ind] = MetalsSNIA[ind]/Density[ind];
                 }
             }
         }
