+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -23,6 +23,7 @@
 	     float *TemperatureUnits, float *TimeUnits,
 	     float *VelocityUnits, float *MassUnits, float Time);
     int transformComovingWithStar(float* Density, float* Metals, 
+        float* MetalsSNII, float* MetalsSNIA,
         float* Vel1, float* Vel2, float* Vel3, 
         float up, float vp, float wp,
         int sizeX, int sizeY, int sizeZ, int direction);
@@ -34,7 +35,8 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
                         float* up, float* vp, float* wp,
                         float* xp, float* yp, float* zp,
                         int ip, int jp, int kp,
-                        int size, float* muField, int winds){
+                        int size, float* muField, int winds, int nSNII,
+                        int nSNIA, float starMetal, int isP3){
     
     /*
      This routine will create an isocahedron of coupling particles, where we determine
@@ -44,11 +46,12 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     */
     bool debug = true;
     bool criticalDebug = true;
+    bool printout = debug && !winds;
     int index = ip+jp*GridDimension[0]+kp*GridDimension[0]*GridDimension[1];
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
     
     /* Compute size (in floats) of the current grid. */
-    float stretchFactor =1.;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
+    float stretchFactor =1.0;//1.5/sin(M_PI/10.0);  // How far should cloud particles be from their host
                                 // in units of dx. Since the cloud forms a sphere shell, stretchFactor > 1 is not recommended
     size = 1;
     for (int dim = 0; dim < GridRank; dim++)
@@ -73,7 +76,7 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     } 
     FLOAT dx = CellWidth[0][0];        
 
-    if (debug)
+    if (printout)
         fprintf(stdout, "depositing quantities: Energy %e, Mass %e, Metals %e\n",
             ejectaEnergy, ejectaMass, ejectaMetal);
 
@@ -81,23 +84,33 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         get metallicity field and set flag; assumed true thoughout feedback
         since so many quantities are metallicity dependent
      */
-    int MetallicityField = FALSE, MetalNum;
+    int MetallicityField = FALSE, MetalNum, MetalIaNum, MetalIINum, SNColourNum;
     if ((MetalNum = FindField(Metallicity, FieldType, NumberOfBaryonFields))
         != -1)
         MetallicityField = TRUE;
-    else
+    else{
+        fprintf(stdout, "MechStars only functions with metallicity field enabled!");
+        ENZO_FAIL("Grid_MechStarsDepositFeedback: 91");
         MetalNum = 0;
-
+    }
+    MetalIaNum = FindField(MetalSNIaDensity, FieldType, NumberOfBaryonFields);
+    MetalIINum = FindField(MetalSNIIDensity, FieldType, NumberOfBaryonFields);
+    SNColourNum = FindField(SNColour, FieldType, NumberOfBaryonFields);
[…]
