### 2019-09-18 #1636 @danielskatz
The next steps will be
1. The authors (@bwoshea) to make any remaining changes (sorry, I can't easily tell if there's more to do or if you've already done this because the PRs didn't reference this review thread) and saying that they are done here.
1. The reviewers (@zingale and @rtfisher) to confirm that they are satisfied by checking off the remaining items in their review checklists above and by saying so in this thread.
