### 2019-09-19 #1636 @bwoshea
@danielskatz  we have made the remaining changes!
