### 2019-09-19 #1636 @danielskatz
👋 @zingale and @rtfisher: please confirm that you are satisfied by checking off the remaining items in your review checklists above and by saying so in this thread.  If not, please say what you think any current issues are.
