### 2019-09-19 #1636 @danielskatz
👋 @bwoshea - to finish up now, I need you to

1. deposit the repository in an archive (e.g. zenodo) and to report the DOI here
1. tell me the latest version number here
1. confirm that the paper text and references are final
