### 2019-09-20 d317031
timing prints and fixed typedefs.h
