### 2019-09-20 149be45
Added scaling graph to documentation
