### 2019-09-22 #1636 @danielskatz
👋 @bwoshea - to finish up now, I need you to

deposit the repository in an archive (e.g. zenodo) and report the DOI here
tell me the latest version number here
confirm that the paper text and references are final
