### 2019-09-23 dc15e5b
Multiple optimization to try and make winds tractable.  Theres just so many goddamn stars!
