+++ src/enzo/Grid_MechStarsFeedbackRoutine.C
@@ -26,13 +26,15 @@
                     float TimeUnits, float dtFixed);
     int determineWinds(float age, float* eWinds, float* zWinds, float* mWinds,
                         float massMsun, float zZsun, float TimeUnits, float dtFixed);
-    int checkCreationCriteria(float* Density, float* Metals,
+int checkCreationCriteria(float* Density, float* Metals,
                         float* Temperature,float* DMField,
-                        float* Vel1, float* Vel2, float* Vel3,
-                        float* CoolingTime, int GridDim,
-                        float* shieldedFraction, float* freeFallTime,
-                        float* dynamicalTime, int i, int j, int k,
-                        float Time, float* RefinementField, float CellWidth);
+                        float* Vel1, float* Vel2, float* Vel3, 
+                        float* CoolingTime, int* GridDim,
+                        float* shieldedFraction, float* freeFallTime, 
+                        float* dynamicalTime, int i, int j, int k, 
+                        float Time, float* RefinementField, float CellWidth,
+                        bool* gridShouldFormStars, bool* notEnoughMetals, 
+                        int continuingFormation, int* seedIndex);
     int FindField(int field, int farray[], int numfields);
     int GetUnits(float *DensityUnits, float *LengthUnits,
 	     float *TemperatureUnits, float *TimeUnits,
@@ -42,14 +44,17 @@
 
 
 
-int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
+int grid::MechStars_FeedbackRoutine(int level, float* mu_field, 
+            float* Temperature, float* CoolingTime, float* DMField)
 {
 
     //fprintf(stdout,"IN FEEDBACK ROUTINE\n  %d   %d   %d\n",
         //SingleSN, StellarWinds, UnrestrictedSN);
     float stretchFactor = 1.0;//1/sqrt(2) to cover cell diagonal
     /* Get units to use */
-    bool debug = false;
+    bool SingleWinds = true; // flag to consolidate wind feedback into one event centered on most massive cell in grid
+            // I wouldn't recommend this for unigrid runs...
+    bool debug = true;
     float startFB = MPI_Wtime();
     int dim, i, j, k, index, size, field, GhostZones = NumberOfGhostZones;
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
@@ -82,15 +87,25 @@ int grid::MechStars_FeedbackRoutine(int level, float* mu_field)
         get metallicity field and set flag; assumed true thoughout feedback
         since so many quantities are metallicity dependent
      */
-    int MetallicityField = FALSE, MetalNum;
+    int MetallicityField = FALSE, MetalNum, SNColourNum=-1;
     if ((MetalNum = FindField(Metallicity, FieldType, NumberOfBaryonFields))
         != -1)
         MetallicityField = TRUE;
     else
         MetalNum = 0;
-
+    if (MechStarsSeedField) 
+        SNColourNum = FindField(SNColour, FieldType, NumberOfBaryonFields);
+    float* totalMetal = new float [size];
+    for (int i = 0; i < size; i++){
+        totalMetal[i] = BaryonField[MetalNum][i];
+        if (MechStarsSeedField)
+            totalMetal[i] += BaryonField[SNColourNum][i];
+    }
     int numSN = 0; // counter of events
     int c = 0; // counter of particles
+    float maxD = 0.0;
+    int maxI=0, maxJ=0, maxK=0;
+    int maxindex=0;
     /* Begin Iteration of all particles */
     // printf("\nIterating all particles  ");
     for (int pIndex=0; pIndex < NumberOfParticles; pIndex++){
[…]
