+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 #include <mpi.h>
+#include <string.h>
 #include "ErrorExceptions.h"
 #include "macros_and_parameters.h"
 #include "typedefs.h"
@@ -25,13 +26,15 @@
     int transformComovingWithStar(float* Density, float* Metals, 
         float* MetalsSNII, float* MetalsSNIA,
         float* Vel1, float* Vel2, float* Vel3, 
+        float* TE, float* GE,
         float up, float vp, float wp,
         int sizeX, int sizeY, int sizeZ, int direction);
     int FindField(int field, int farray[], int numfields);
 
 
 int grid::MechStars_DepositFeedback(float ejectaEnergy, 
-                        float ejectaMass, float ejectaMetal,
+                        float ejectaMass, float ejectaMetal, 
+                        float* totalMetals,
                         float* up, float* vp, float* wp,
                         float* xp, float* yp, float* zp,
                         int ip, int jp, int kp,
@@ -105,33 +108,38 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
 
     /* Make copys of fields to work with. These will conatin the added deposition
         of all quantities and be coupled to the grid after the cic deposition. */
-    float totalMetals [size];
-    float density [size];
-    float metals [size];
-    float metalsII [size];
-    float metalsIA [size];
-    float metalsIII [size];
-    float u [size];
-    float v [size];
-    float w [size];
-    float totalEnergy  [size];
-    float gasEnergy [size];
-    for (int i=0; i<size; i++){
-        density[i] = 0.0;
-        metals[i] = 0.0;
-        metalsII [i] = 0.0;
-        metalsIA[i] = 0.0;
-        metalsIII[i] = 0.0;
-        u[i] = 0.0;
-        v[i] =0.0;
-        w[i] = 0.0;
-        totalEnergy[i] = 0.0;
-        gasEnergy[i] = 0.0;
-        totalMetals[i] = BaryonField[MetalNum][i];
-        // if (StarMakerTypeIaSNe) totalMetals[i] += BaryonField[MetalIaNum][i]*BaryonField[DensNum][i];
-        // if (StarMakerTypeIISNeMetalField) totalMetals[i] += BaryonField[MetalIINum][i]*BaryonField[DensNum][i];
-        if (PopIIISupernovaUseColour) totalMetals[i] += BaryonField[SNColourNum][i];
-    }
+    float* density = BaryonField[DensNum];
+    float* metals = BaryonField[MetalNum];
+    float* metalsII = BaryonField[MetalIINum];
+    float* metalsIA = BaryonField[MetalIaNum];
+    float* metalsIII =BaryonField[SNColourNum];
+    float* u = BaryonField[Vel1Num];
+    float* v = BaryonField[Vel2Num];
+    float* w = BaryonField[Vel3Num];
+    float* totalEnergy = BaryonField[TENum];
+    float* gasEnergy = BaryonField[GENum];
+    // memset(density, 0, size*sizeof(float));
+    // memset(metals, 0, size*sizeof(float));
+    // memset(metalsII, 0, size*sizeof(float));
+    // memset(metalsIA, 0, size*sizeof(float));
+    // memset(metalsIII, 0, size*sizeof(float));
+    // memset(u, 0, size*sizeof(float));
+    // memset(v, 0, size*sizeof(float));
+    // memset(w, 0, size*sizeof(float));
+    // memset(totalEnergy, 0, size*sizeof(float));
+    // memset(gasEnergy, 0, size*sizeof(float));
+    // for (int i=0; i<size; i++){
+    //     density[i] = 0.0;
[…]

[…]
