### 2019-09-23 d825fab
Refinements to continual formation routine in Grid_MechStarsDepositFeedback.C
