### 2019-09-23 a239f47
forgot to account for metals in continual formation!
