### 2019-09-23 fffc0c9
Removed modification of SNII and SNIa fields at particle creation- these are fractions and dont need modification
