### 2019-09-23 a03b77e
Added timing prints to star creation and accretion printout
