### 2019-09-23 7cef58a
Moved temperature calculation to calling routines and just pass pointers
