### 2019-09-23 6acf305
Again, moved total metals calculation to calling routines to save time in MechStars routines
