### 2019-09-25 8acef27
Bugfixes from merge with enzo-dev issues. Mostly compile time syntax errors and merge goofs. Couple updates. Updating constants in my own code to header constants
