+++ src/enzo/Grid_DeleteAllButParticles.C
@@ -30,8 +30,10 @@ void grid::DeleteAllButParticles()
   int i;
  
   //  this->DeleteParticles();
+#ifdef INDIVIDUALSTAR
   this->DeleteStellarAbundances();
- 
+#endif
+
   for (i = 0; i < MAX_DIMENSION; i++) {
     delete [] ParticleAcceleration[i];
     delete [] AccelerationField[i];
