+++ src/enzo/global_data.h
@@ -1268,9 +1268,5 @@ EXTERN int StellarYieldsScaledSolarInitialAbundances;
 EXTERN int ResetStellarAbundances;
 
 
-EXTERN int   BoundaryMassFluxFieldNumbers[MAX_NUMBER_OF_BARYON_FIELDS];
-EXTERN float BoundaryMassFluxContainer[MAX_NUMBER_OF_BARYON_FIELDS];
-EXTERN char *BoundaryMassFluxFilename;
-
 #endif
 
