### 2019-09-25 82fd14e
Star Particle object communication buffer fixes circa 2016 that fell through the cracks in getting pulled into enzo-dev. (all credit John Wise)
