+++ src/enzo/macros_and_parameters.h
@@ -369,6 +369,8 @@ typedef long long int   HDF5_hid_t;
 #define POW(X,Y) pow((double) (X), (double) (Y))
 #define COS(X) cos((double) (X))
 #define SIN(X) sin((double) (X))
+#define ceil_log2(size) ((size) > 1 ? ((size_t)pow(2, (int)log2(size-1)+1)) : 1 )
+
 #ifdef CONFIG_PFLOAT_4
 #define MODF(X,Y) modff((X), (Y))
 #elif CONFIG_PFLOAT_8
