### 2019-09-26 b9d664a
missed checks on metal field flags to prevent access when not using SNII, SNIa and PIII metals
