+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -47,10 +47,12 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         and all have radius dx from the source particle. 
         Each vertex particle will then be CIC deposited to the grid!
     */
+    printf("In Feedback deposition\n");
     bool debug = false;
     bool criticalDebug = false;
     bool printout = debug && !winds;
     int index = ip+jp*GridDimension[0]+kp*GridDimension[0]*GridDimension[1];
+    if (printout) printf("Host index = %d", index);
     int DensNum, GENum, TENum, Vel1Num, Vel2Num, Vel3Num;
     
     /* Compute size (in floats) of the current grid. */
@@ -80,14 +82,14 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     FLOAT dx = CellWidth[0][0];        
 
     if (printout)
-        fprintf(stdout, "depositing quantities: Energy %e, Mass %e, Metals %e\n",
+        printf("depositing quantities: Energy %e, Mass %e, Metals %e\n",
             ejectaEnergy, ejectaMass, ejectaMetal);
 
     /* 
         get metallicity field and set flag; assumed true thoughout feedback
         since so many quantities are metallicity dependent
      */
-    int MetallicityField = FALSE, MetalNum, MetalIaNum, MetalIINum, SNColourNum;
+    int MetallicityField = FALSE, MetalNum, MetalIaNum=-1, MetalIINum=-1, SNColourNum=-1;
     if ((MetalNum = FindField(Metallicity, FieldType, NumberOfBaryonFields))
         != -1)
         MetallicityField = TRUE;
@@ -96,16 +98,20 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         ENZO_FAIL("Grid_MechStarsDepositFeedback: 91");
         MetalNum = 0;
     }
-    MetalIaNum = FindField(MetalSNIaDensity, FieldType, NumberOfBaryonFields);
-    MetalIINum = FindField(MetalSNIIDensity, FieldType, NumberOfBaryonFields);
-    SNColourNum = FindField(SNColour, FieldType, NumberOfBaryonFields);
-    if (PopIIISupernovaUseColour && SNColourNum==-1) ENZO_FAIL("Cant use Seed Field without SNColour field");
+    if(StarMakerTypeIaSNe)
+        MetalIaNum = FindField(MetalSNIaDensity, FieldType, NumberOfBaryonFields);
+    if (StarMakerTypeIISNeMetalField)
+        MetalIINum = FindField(MetalSNIIDensity, FieldType, NumberOfBaryonFields);
+    if (MechStarsSeedField)
+        SNColourNum = FindField(SNColour, FieldType, NumberOfBaryonFields);
+    
+    if (MechStarsSeedField && SNColourNum<0) ENZO_FAIL("Cant use Seed Field without SNColour field");
     /* set other units that we need */
     MassUnits = DensityUnits*pow(LengthUnits*dx, 3)/SolarMass; //Msun!
     float EnergyUnits = DensityUnits*pow(LengthUnits*dx, 3) 
                     * VelocityUnits*VelocityUnits;//[g cm^2/s^2] -> code_energy
     float MomentaUnits = MassUnits*VelocityUnits;  
-
+    if (printout) printf("Making pointer copies\n");
     /* Make copys of fields to work with. These will conatin the added deposition
         of all quantities and be coupled to the grid after the cic deposition. */
     float* density = BaryonField[DensNum];
@@ -151,10 +157,10 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
 
     
[…]
