### 2019-09-30 619f7cc
general debugging, implemented maximum velocity for created star particles.
