### 2019-10-01 fd8d8b1
first commit to incorporate rad hydro with mechanical star maker
