+++ src/enzo/Grid_MechStarsDepositFeedback.C
@@ -124,28 +124,7 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     float* w = BaryonField[Vel3Num];
     float* totalEnergy = BaryonField[TENum];
     float* gasEnergy = BaryonField[GENum];
-    // memset(density, 0, size*sizeof(float));
-    // memset(metals, 0, size*sizeof(float));
-    // memset(metalsII, 0, size*sizeof(float));
-    // memset(metalsIA, 0, size*sizeof(float));
-    // memset(metalsIII, 0, size*sizeof(float));
-    // memset(u, 0, size*sizeof(float));
-    // memset(v, 0, size*sizeof(float));
-    // memset(w, 0, size*sizeof(float));
-    // memset(totalEnergy, 0, size*sizeof(float));
-    // memset(gasEnergy, 0, size*sizeof(float));
-    // for (int i=0; i<size; i++){
-    //     density[i] = 0.0;
-    //     metals[i] = 0.0;
-    //     metalsII [i] = 0.0;
-    //     metalsIA[i] = 0.0;
-    //     metalsIII[i] = 0.0;
-    //     u[i] = 0.0;
-    //     v[i] =0.0;
-    //     w[i] = 0.0;
-    //     totalEnergy[i] = 0.0;
-    //     gasEnergy[i] = 0.0;
-    // }
+
     /* Transform coordinates so that metals is fraction (rho metal/rho baryon)
         u, v, w -> respective momenta.  Use -1 to reverse transform after.*/
     /* these temp arrays are implicitly comoving with the star! */
@@ -236,9 +215,6 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
         zmean += totalMetals[index+GridDimension[0]*ind]*BaryonField[DensNum][index+GridDimension[0]*ind];
         zmean += totalMetals[index+GridDimension[0]*GridDimension[1]*ind]*BaryonField[DensNum][index+GridDimension[0]*GridDimension[1]*ind];
 
-        // if (printout) fprintf(stdout, "MuField = %f %f %f\nmax = %d ; %d %d %d\n",
-        //     muField[index+ind], muField[index+GridDimension[0]*ind], muField[index+ind*GridDimension[0]*GridDimension[1]], GridDimension[0]*GridDimension[1]*GridDimension[2],
-        //     index+ind, index+GridDimension[0]*ind, index+ind*GridDimension[0]*GridDimension[1]);
         nmean += BaryonField[DensNum][index+ind]*BaryonField[DensNum][index+ind]*DensityUnits/mh/max(0.6,muField[index+ind]);
         nmean += BaryonField[DensNum][index+GridDimension[0]*ind]*
             BaryonField[DensNum][index+GridDimension[0]*ind]*DensityUnits/mh/max(muField[index+GridDimension[0]*ind],0.6);
@@ -272,33 +248,40 @@ int grid::MechStars_DepositFeedback(float ejectaEnergy,
     float fz = (zZsun < 0.01)? (2.0): (pow(zZsun, -0.14));
 
     /* Cooling radius as in Hopkins, but as an average over cells */
+    
     float CoolingRadius = 28.4 *
         pow(max(0.001,nmean), -3.0/7.0)
         *pow(ejectaEnergy/1.0e51, 2.0/7.0)* fz;
     if (printout)fprintf(stdout, "cooling radius [pc] = %e\n %f %e %f %e %e \n", 
             CoolingRadius, nmean, ejectaEnergy/1e51, fz, zmean, dmean);
[…]
