+++ src/enzo/StarParticleData.h
@@ -162,5 +162,6 @@ SPEXTERN int AnalyticSNRShellMass;
 SPEXTERN int UnrestrictedSN;
 SPEXTERN int MechStarsSeedField;
 SPEXTERN float MechStarsCriticalMetallicity;
+SPEXTERN int MechStarsRadiationSpectrum;
 
 #endif
