### 2019-10-02 #1636 @bwoshea
@danielskatz sorry for the delay.  We now have a DOI:  10.5281/zenodo.3468822:

 <a href="https://zenodo.org/badge/latestdoi/178911062"><img src="https://zenodo.org/badge/178911062.svg" alt="DOI"></a>

The lates version number is Enzo v2.6.1.

We are verifying the paper text and references right now.  The one issue I can see with the paper is the overlapping text in the bottom-left corner of the first page, where the 'LICENSE' block overlaps the footer:

![JOSS_paper_issue](https://user-images.githubusercontent.com/7969676/66050071-81abfd00-e4fa-11e9-87ed-3fad5d01ef1f.jpeg)

I anticipate that I will get back to you by no later than tomorrow finalizing the paper text and references.
