### 2019-10-02 #1636 @arfon
> We are verifying the paper text and references right now. The one issue I can see with the paper is the overlapping text in the bottom-left corner of the first page, where the 'LICENSE' block overlaps the footer:

@bwoshea @danielskatz - this is an issue with our LaTeX template not handling long author lists very well. @danielskatz - feel free to accept this paper with Whedon here but I'll need to manually update the PDF afterwards to fix the layout issues.
