### 2019-10-02 9172a79
turning cooling/heating back on in the midplane
