### 2019-10-03 #1636 @bwoshea
@danielskatz We've proofread the paper and pushed a small change.  The revised PDF looks good to me.  I think we're good to go!
