### 2019-10-03 #1636 @danielskatz
Actually, if you want to take over given that you have to generate the PDF manually for the spacing issue, please go ahead - this is fully ready to publish
